Separate bin from lib
