+++ tools/tenderly/package.json
@@ -7,12 +7,12 @@
   "license": "MIT",
   "module": "index.mjs",
   "bin": {
-    "tenderly-fork": "./fork.mjs",
-    "tenderly-geteth": "./geteth.mjs",
-    "tenderly-getsnx": "./getsnx.mjs",
-    "tenderly-load": "./load.mjs",
-    "tenderly-save": "./save.mjs",
-    "tenderly-unfork": "./unfork.mjs"
+    "tenderly-fork": "./bin/tenderly-fork.mjs",
+    "tenderly-geteth": "./bin/tenderly-geteth.mjs",
+    "tenderly-getsnx": "./bin/tenderly-getsnx.mjs",
+    "tenderly-load": "./bin/tenderly-load.mjs",
+    "tenderly-save": "./bin/tenderly-save.mjs",
+    "tenderly-unfork": "./bin/tenderly-unfork.mjs"
   },
   "dependencies": {
     "dotenv": "^16.0.0",
